requirements.txt
flask==2.3.2
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
PyJWT==2.8.0
cryptography==41.0.3